Move otconfig parser to own module
